ReportCardEntryDao can now return emails addresses of attendees in given labwork
